Giant refactor (WIP)
